+++ Engine/Graphics/Window/glfw_window.py
@@ -12,6 +12,7 @@ from ...Control.keyboard import Keyboard
 from ...Control.mouse import Mouse
 from ..GUI.window import _drawText, SimpleButton, textInput, _drawTextBox, bgframe
 from ..GUI.imgui import imguiInit, imguiInputs, imguiRender
+from ..GUI.console import Console
 from .console import consoleHandler
 
 #_drawText
@@ -60,8 +61,7 @@ class Window:
         self.render_type = render_type
 
         self.console = consoleHandler(self)
-        self.console_input = textInput(Vec2(0, 240), Vec2(235, 12), Color3(1, 1, 1))
-        self.console_button = SimpleButton("Send", Vec2(240, 240), Vec2(40, 12), Color3(1, 1, 1), self.consoleCommand)
+        self.console_gui = Console(self.console)
 
         self.fullscreen = False
         self.fullscreen_switching = True
@@ -343,7 +343,6 @@ class Window:
         self.console_input.setValue("")
         self.console.handleCommand(str(cmd))
 
-
     def _render_frame(self, update=None):
         
         # GET CURRENT WINDOW SIZES
@@ -501,13 +500,6 @@ class Window:
                     self.drawText(label, Vec2(0, index*padding), debug_only=True)
             elif self.debugmenu == 2 and debug:
                 self.drawText(f"FPS: {self.fps}", Vec2(0, 0), debug_only=True)
-            elif self.debugmenu == 3 and debug:
-                bgframe(Vec2(0, 0), Vec2(280, 228), Color4(0, 0, 0, -0.5))
-                self.drawTextBox(self.console.output, charslen=46, debug_only=True, color=Color3(1, 1, 1))
-                self.console_input._process(self)
-                self.console_input._draw(self)
-                self.console_button._process(self)
-                self.console_button._draw(self)
 
         # FULL SCREEN
 
@@ -520,10 +512,15 @@ class Window:
         if self.debugger:
             self.debugger._debugger_work()
         
+        # CONSOLE
+
+        if self.debugmenu == 3 and debug:
+            self.console_gui.drawConsole()
+
         # IMGUI RENDER
 
         imguiRender(self.impl)
-
+        
         # WINDOW PROCESS
         
         glfw.swap_buffers(self.window)