+++ Engine/Misc/idgenerator.py
@@ -1,0 +1,7 @@
+from ..Kernel.modules import time
+
+_counter = -1
+
+def generateID():
+    _counter += 1
+    return int(time.time()) + _counter