+++ CHANGELOG.md
@@ -147,4 +147,4 @@ added **GUIArrowButton**
 
 # 04.08.26
 
-updated **GUI**
+updated **imGUI** + **Console GUI**