+++ CHANGELOG.md
@@ -148,3 +148,5 @@ added **GUIArrowButton**
 # 04.08.26
 
 updated **imGUI** + **Console GUI**
+
+added **rayCast**