+++ CHANGELOG.md
@@ -149,4 +149,4 @@ added **GUIArrowButton**
 
 updated **imGUI** + **Console GUI** + **Debug GUI**
 
-added **rayCast**
+added **rayCast** + **generateID**